Worked on KSplash and KDM configuration
With all the different resolutions calculated by /script/update-desktop-theme, we can safely assume, that KSPlash won't distort the login picture any more. To have a clean transition from KDM to Ksplash we had formerly set the scale mode of KDM to "scale" which means KDM would distort the picture too. Now it should be safe to set it back to "crop".